dev: extract MockHelix to own file
